添加post示例

+++ src/main/java/com/quanqinle/myworld/dao/TaxRateRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface TaxRateRepository extends JpaRepository<TaxRate, Integer> {
-    List<TaxRate> findByRangeLowestLessThanOrderByRangeLowestDesc(float income);
+    List<TaxRate> findByRangeLowestLessThanOrderByRangeLowestDesc(double income);
 }
