+++ src/main/java/com/quanqinle/myworld/service/impl/TaxRateServiceImpl.java
@@ -33,7 +33,7 @@ public class TaxRateServiceImpl implements TaxRateService {
 
 	@Override
 	@Cacheable(cacheNames = {"tax", "taxrate"})
-	public TaxRate getTaxRateByRange(float income) {
+	public TaxRate getTaxRateByRange(double income) {
 //        return null;
 		if (income <= 0) {
 			return null;
